# .cloned()'ed