--- a/src/branch.rs
+++ b/src/branch.rs
@@ -124,7 +124,8 @@ macro_rules! alt (
 
   (__impl $i:expr, $subrule:ident!( $($args:tt)*) | $($rest:tt)*) => (
     {
-      let res = $subrule!($i, $($args)*);
+      let i_ = $i.clone();
+      let res = $subrule!(i_, $($args)*);
       match res {
         $crate::IResult::Done(_,_)     => res,
         $crate::IResult::Incomplete(_) => res,
@@ -135,7 +136,8 @@ macro_rules! alt (
 
   (__impl $i:expr, $subrule:ident!( $($args:tt)* ) => { $gen:expr } | $($rest:tt)+) => (
     {
-      match $subrule!( $i, $($args)* ) {
+      let i_ = $i.clone();
+      match $subrule!(i_, $($args)* ) {
         $crate::IResult::Done(i,o)     => $crate::IResult::Done(i,$gen(o)),
         $crate::IResult::Incomplete(x) => $crate::IResult::Incomplete(x),
         $crate::IResult::Error(_)      => {
@@ -155,7 +157,8 @@ macro_rules! alt (
 
   (__impl $i:expr, $subrule:ident!( $($args:tt)* ) => { $gen:expr }) => (
     {
-      match $subrule!( $i, $($args)* ) {
+      let i_ = $i.clone();
+      match $subrule!(i_, $($args)* ) {
         $crate::IResult::Done(i,o)     => $crate::IResult::Done(i,$gen(o)),
         $crate::IResult::Incomplete(x) => $crate::IResult::Incomplete(x),
         $crate::IResult::Error(_)      => {
@@ -171,7 +174,8 @@ macro_rules! alt (
 
   (__impl $i:expr, $subrule:ident!( $($args:tt)*)) => (
     {
-      match $subrule!( $i, $($args)* ) {
+      let i_ = $i.clone();
+      match $subrule!(i_, $($args)* ) {
         $crate::IResult::Done(i,o)     => $crate::IResult::Done(i,o),
         $crate::IResult::Incomplete(x) => $crate::IResult::Incomplete(x),
         $crate::IResult::Error(_)      => {
@@ -207,7 +211,8 @@ macro_rules! alt_complete (
 
   ($i:expr, $subrule:ident!( $($args:tt)*) | $($rest:tt)*) => (
     {
-      let res = complete!($i, $subrule!($($args)*));
+      let i_ = $i.clone();
+      let res = complete!(i_, $subrule!($($args)*));
       match res {
         $crate::IResult::Done(_,_) => res,
         _ => alt_complete!($i, $($rest)*),
@@ -217,7 +222,8 @@ macro_rules! alt_complete (
 
   ($i:expr, $subrule:ident!( $($args:tt)* ) => { $gen:expr } | $($rest:tt)+) => (
     {
-      match complete!($i, $subrule!($($args)*)) {
+      let i_ = $i.clone();
+      match complete!(i_, $subrule!($($args)*)) {
         $crate::IResult::Done(i,o) => $crate::IResult::Done(i,$gen(o)),
         _ => alt_complete!($i, $($rest)*),
       }
@@ -328,7 +334,8 @@ macro_rules! alt_complete (
 macro_rules! switch (
   (__impl $i:expr, $submac:ident!( $($args:tt)* ), $($p:pat => $subrule:ident!( $($args2:tt)* ))|* ) => (
     {
-      match map!($i, $submac!($($args)*), |o| Some(o)) {
+      let i_ = $i.clone();
+      match map!(i_, $submac!($($args)*), |o| Some(o)) {
         $crate::IResult::Error(e)      => $crate::IResult::Error(error_node_position!(
             $crate::ErrorKind::Switch, $i, e
         )),