allows to use non-Copy types with most parsers (by cloning the input at all relevant places)