@s3bk 2017-04-26
# Benchmark check
## result: no changes outside std-deviation

## before
```
     Running target/release/deps/http-21f389d5e78fa41b

running 1 test
test one_test ... bench:       2,423 ns/iter (+/- 65)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured

     Running target/release/deps/ini-c9ae1c033ebfff90

running 10 tests
test parse_category_test ... ignored
test parse_category_then_multiple_keys_and_values_test ... ignored
test parse_key_value_test ... ignored
test parse_key_value_with_comment_test ... ignored
test parse_key_value_with_space_test ... ignored
test parse_multiple_categories_test ... ignored
test parse_multiple_keys_and_values_test ... ignored
test bench_ini                 ... bench:       2,697 ns/iter (+/- 64) = 40 MB/s
test bench_ini_key_value       ... bench:         147 ns/iter (+/- 81) = 122 MB/s
test bench_ini_keys_and_values ... bench:         231 ns/iter (+/- 3) = 190 MB/s

test result: ok. 0 passed; 0 failed; 7 ignored; 3 measured

     Running target/release/deps/ini_str-7c817276147ec404

running 8 tests
test parse_category_test ... ignored
test parse_category_then_multiple_keys_and_values_test ... ignored
test parse_key_value_test ... ignored
test parse_key_value_with_comment_test ... ignored
test parse_key_value_with_space_test ... ignored
test parse_multiple_categories_test ... ignored
test parse_multiple_keys_and_values_test ... ignored
test bench_ini_str ... bench:       4,074 ns/iter (+/- 159) = 26 MB/s

test result: ok. 0 passed; 0 failed; 7 ignored; 1 measured

     Running target/release/deps/json-927b4aaf8a824a78

running 1 test
test json_bench ... bench:       8,503 ns/iter (+/- 202)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured
```

## after
```
     Running target/release/deps/http-4462e31d4bddc0ae

running 1 test
test one_test ... bench:       2,462 ns/iter (+/- 136)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured

     Running target/release/deps/ini-4ff7f590412ac43a

running 10 tests
test parse_category_test ... ignored
test parse_category_then_multiple_keys_and_values_test ... ignored
test parse_key_value_test ... ignored
test parse_key_value_with_comment_test ... ignored
test parse_key_value_with_space_test ... ignored
test parse_multiple_categories_test ... ignored
test parse_multiple_keys_and_values_test ... ignored
test bench_ini                 ... bench:       2,662 ns/iter (+/- 146) = 40 MB/s
test bench_ini_key_value       ... bench:         144 ns/iter (+/- 3) = 125 MB/s
test bench_ini_keys_and_values ... bench:         235 ns/iter (+/- 7) = 187 MB/s

test result: ok. 0 passed; 0 failed; 7 ignored; 3 measured

     Running target/release/deps/ini_str-c11befa65d3b4bbc

running 8 tests
test parse_category_test ... ignored
test parse_category_then_multiple_keys_and_values_test ... ignored
test parse_key_value_test ... ignored
test parse_key_value_with_comment_test ... ignored
test parse_key_value_with_space_test ... ignored
test parse_multiple_categories_test ... ignored
test parse_multiple_keys_and_values_test ... ignored
test bench_ini_str ... bench:       4,144 ns/iter (+/- 190) = 26 MB/s

test result: ok. 0 passed; 0 failed; 7 ignored; 1 measured

     Running target/release/deps/json-ae53913aa97b9c14

running 1 test
test json_bench ... bench:       8,389 ns/iter (+/- 142)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured
```