@s3bk 2017-05-08
PR updated to include the latest changes